remove up/down commands
